% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_signature_vs_dataset function.R
\name{check_HR2_signature_vs_dataset}
\alias{check_HR2_signature_vs_dataset}
\title{Validity check in gene signatures and gene expression datasets}
\usage{
check_HR2_signature_vs_dataset(data_norm, sig_up_df, sig_dn_df)
}
\arguments{
\item{data_norm}{Normalized Gene expression data matrix}

\item{sig_up_df}{Up-regulated gene-set}

\item{sig_dn_df}{Down-regulated gene-set}
}
\value{
Filtered Normalized Gene expression data matrix
}
\description{
performs validity check against gene signatures and gene expression dataset and
filters gene expression dataset accordingly
}
\examples{
check_signature_vs_dataset(data_norm, sig_up_df, sig_dn_d)
}
\author{
Yi-Hsuan Lee \email{yi-hsuan.lee@cranfield.ac.uk}
}
