% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{check_matrix_sig_overlap}
\alias{check_matrix_sig_overlap}
\title{Checking overlap between matrix and signature gene symbols}
\usage{
check_matrix_sig_overlap(input, sig_df)
}
\arguments{
\item{input}{The expression matrix after being processed and
returned by read_input_file function}

\item{sig_df}{The signature data frame after being
processed and returned by read_signature_data function}
}
\value{
A Venn diagram representing the overlap of
gene symbols between the two files
}
\description{
Plots a Venn diagram checking which gene symbols
overlap between the signature file
and the expression matrix file.
The user gets an estimates of how many gene symbols overlap
between the two files.
}
\examples{
check_matrix_sig_overlap(input, sig_df)
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
