% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_pre-processing.R
\name{check_signature_vs_dataset}
\alias{check_signature_vs_dataset}
\title{Validity check in gene signatures and gene expression datasets}
\usage{
check_signature_vs_dataset(norm_data, sig_df)
}
\arguments{
\item{norm_data}{Normalized gene expression data matrix}

\item{sig_df}{signature gene-set}
}
\value{
Filtered Normalized Gene expression data matrix
}
\description{
This function performs a validity check against gene signatures and gene expression dataset and
filters gene expression dataset accordingly.
}
\examples{
\dontrun{check_signature_vs_dataset(norm_data, sig_df)}
}
\author{
Yi-Hsuan Lee \email{yi-hsuan.lee@cranfield.ac.uk}
}
