% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify}
\alias{classify}
\title{Sample classification based on gene signature expression}
\usage{
classify(sig_df, data_se, up_thresh = NULL, dn_thresh = NULL)
}
\arguments{
\item{sig_df}{Gene expression signature for a specific pathway given as data frame
containing a list of genes that are the most differentially expressed when
a specific pathway is active. Up-regulated genes are given an
expression of 1, while down-regulated genes are given an expression of -1.}

\item{data_se}{Expression data set matrix given as a matrix or data frame
comprising the expression levels of genes in a set of samples. Gene expression matrices can be from RNASeq
or microarray transcriptomic data sets.}

\item{up_thresh}{Numerical vector of thresholds for the up-regulated gene set between -1 and 1.}

\item{dn_thresh}{Numerical vector of thresholds for the down-regulated gene set between -1 and 1.}
}
\value{
A data frame containing a list of samples and their classified pathway activity
(Active, Inactive or Uncertain).
}
\description{
Classifies samples according to pathway activity using Gene Set Variation
Analysis (GSVA)
}
\examples{
# Default thresholds for up-regulated and down-regulated gene-sets
classify(HER2_sig_df, HER_data_se)
# up-regulated gene set threshold: bottom = -0.4, top = 0.4
# down-regulated gene set thresholds: bottom = -0.4, top = 0.4
classify(HER2_sig_df, HER_data_se, up_thresh=c(-0.4, 0.4),
dn_thresh=c(-0.4, 0.4))
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
