% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_GSVA_abs}
\alias{classify_GSVA_abs}
\title{Classification Using Absolute GSVA Score Thresholds}
\usage{
classify_GSVA_abs(
  sig_df,
  data_se,
  up_thresh.low,
  up_thresh.high,
  dn_thresh.low,
  dn_thresh.high
)
}
\arguments{
\item{sig_df}{Gene expression signature for a specific pathway given as data
frame containing a list of genes that are the most differentially expressed
when a specific pathway is active. Up-regulated genes are given an
expression of 1, while down-regulated genes are given an expression of -1.}

\item{data_se}{Expression data set matrix given as a matrix or data frame
comprising the expression levels of genes for each sample in a data set.
Gene expression matrices can be from RNASeq or microarray transcriptomic data
sets.}

\item{up_thresh.low}{Number denoting the absolute GSVA score threshold for the
categorizing a sample as inconsistent expression with the up-regulated
gene-set from the gene signature.}

\item{up_thresh.high}{Number denoting the absolute GSVA score threshold for
categorizing a sample as having consistent expression with the up-regulated
gene set from the gene signature.}

\item{dn_thresh.low}{Number denoting the absolute GSVA score threshold for
the categorizing a sample as consistent expression with the down-regulated
gene-set from the gene signature.}

\item{dn_thresh.high}{Number denoting the absolute GSVA score threshold for
the categorizing a sample as inconsistent expression with the down-regulated
gene-set from the gene signature.}
}
\value{
A data frame containing a list of samples and their classified
pathway activity (Active, Inactive or Uncertain).
}
\description{
Classifies samples according to pathway activity using the GSVA
algorithm with absolute GSVA scores thresholds for assessing expression
consistency with the up-regulated  and down-regulated gene-sets of the gene
signature.
}
\details{
The absolute threshold GSVA function classifies samples by pathway
activity using
}
\examples{
# Default thresholds for up-regulated and down-regulated gene-sets
classes_df <- classify_GSVA_abs(sig_df, data_se, up_thresh.low=-0.25,
     up_thresh.high=0.25, dn_thresh.low=-0.25, dn_thresh.high=0.35)
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
