% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{.run_GSVA}
\alias{.run_GSVA}
\title{Performs GSVA algorithm using both up-regulated and down-regulated gene-sets
of the gene signature}
\usage{
.run_GSVA(sig_df, data_se)
}
\arguments{
\item{sig_df}{Gene expression signature for a specific pathway given as data
frame containing a list of genes that are the most differentially expressed
when a specific pathway is active. Up-regulated genes are given an
expression of 1, while down-regulated genes are given an expression of -1.}

\item{data_se}{Normalized expression data set matrix given as a matrix or
data frame comprising the expression levels of genes for each sample in a
data set. Gene expression matrices can be from RNASeq or microarray
transcriptomic data sets.}
}
\value{
GSVA scores in the form of a list, containing two data frames: GSVA
scores for up-regulated gene-set for each sample, and GSVA scores for the
down-regulated gene-set GSVA for each sample
}
\description{
Generates GSVA scores for each sample generated by running GSVA
algorithm on the two gene-sets of the gene signature provided.
}
\examples{
run_GSVA(ER_sig_df, ER_data_se1)
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
\keyword{internal}
