% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ER_SET_sig}
\alias{ER_SET_sig}
\title{ER gene signature}
\format{
A data frame with 160 rows (gene names) and 2 variables:
\describe{
\item{gene}{The HUGO gene symbols of a highly differentially expressed gene
when ER pathway is active.}
\item{expression}{The change in expression of the gene when the pathway
is active i.e. -1 for down-regulated genes and 1 for up-regulated genes.}
}
}
\source{
SET index defined by Symmans et al. 2010 (see
\url{https://pubmed.ncbi.nlm.nih.gov/20697068/}).
}
\usage{
ER_SET_sig
}
\description{
A list of up-regulated and down-regulated genes constituting the gene
signature of ER pathway activation. The ER signature was obtained from the
sensitivity to endocrine therapy (SET) genomic index defined using a list of
genes co-expressed with the estrogen receptor (ESR1 or ER) from microarray
data (437 profiles) originating from newly diagnosed breast cancer
independent of outcome and treatment (Symmans et al., 2010).
}
\references{
Symmans, W.F., Hatzis, C., Sotiriou, C., Andre, F., Peintinger,
F., Regitnig, P., Daxenbichler, G., Desmedt, C., Domont, J., Marth, C. and
Delaloge, S., 2010. Genomic index of sensitivity to endocrine therapy for
breast cancer. Journal of clinical oncology, 28(27), p.4111. doi:
\url{https://doi.org/10.1200/jco.2010.28.4273}
}
\keyword{datasets}
