% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ER_TCGA_RNA}
\alias{ER_TCGA_RNA}
\title{ER RNA-seq gene expression data set from TCGA}
\format{
A matrix with 20,124 HUGO gene names (rows) and 60 TCGA barcodes which act as
IDs for each tumour sample (columns).
}
\source{
The Cancer Genome Atlas (TCGA) \url{https://portal.gdc.cancer.gov}).
}
\usage{
ER_TCGA_RNA
}
\description{
A gene expression matrix containing RNA-seq raw read counts for 60 primary
breast cancer tumours (30 tumours of which were estrogen receptor (ER)
positive and 30 of which were ER negative). This data set is a subset of a
much larger data set collected from TCGA containing gene expression data for
1,101 primary breast cancer tumours.
}
\keyword{datasets}
