% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{evaluatematrix}
\alias{evaluatematrix}
\title{Evaluation of gene expression data}
\usage{
evaluatematrix(expdata)
}
\arguments{
\item{expdata}{Gene exoression data matrix or dataframe}
}
\value{
Information about the matrix such as dimensions, range, mean and plots before and after normalization boxplots of matrix
}
\description{
Checks matrix characteristics and normalizes matrix with log cpm normalization
}
\examples{
evaluatematrix(expdata)
}
\author{
Tania Pal \email{taniya.pal.094@cranfield.ac.uk}
}
