% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{gsva_scores_dist}
\alias{gsva_scores_dist}
\title{GSVA score density distribution plot}
\usage{
gsva_scores_dist(expr_mat, sig_df)
}
\arguments{
\item{expr_mat}{Normalised expression data set matrix comprising the
expression levels of genes (rows) for each sample (columns) in a data set. Row names
are gene symbols and column names are sample IDs / names. Gene expression
matrices can contain normalised (logCPM transformed) RNASeq or microarray
transcriptomic data.}

\item{sig_df}{Gene expression signature for a specific pathway given as data
frame with the first column named "gene" containing a list of genes that are
the most differentially expressed when the given pathway is active, and the
second column named "expression" containing their corresponding expression in
the gene signature: -1 for down-regulated genes and 1 for up-regulated genes.}
}
\value{
A density plot displaying distribution of GSVA scores obtained for
the samples using up-regulated and down-regulated gene-sets from the gene
signature
}
\description{
Plots GSVA scores distribution across samples after performing
GSVA algorithm on samples in a dataset (gene expression matrix) for the
up-regulated and down-regulated gene-sets of a given gene expression
signature.
}
\examples{
\dontrun{gsva_scores_dist(ER_dataset, ER_sig)}
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
