% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HER2_data_se1}
\alias{HER2_data_se1}
\title{HER2 gene expression data set 1}
\format{
A data frame with 20,124 rows and 60 variables:
\describe{
\item{gene}{The name of the gene for which expression data is provided.}
\item{sample}{The sample or case ID.}
}
}
\source{
TCGA \url{https://portal.gdc.cancer.gov}).
}
\usage{
HER2_data_se1
}
\description{
A gene expression matrix containing raw read counts from RNA-seq for 60
samples (30 HER2 positive and 30 HER2 negative). The Raw RNA-seq counts data
were obtained from TCGA. The data set is a subset of a larger data set of
filtered for breast cancer-primary tumors from 1,101 samples collected from
TCGA.
}
\keyword{datasets}
