% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HER2_data_se2}
\alias{HER2_data_se2}
\title{HER gene expression data set 2}
\format{
A data frame with 23,113 rows and 20 variables:
\describe{
\item{gene}{The name of the gene for which expression data is provided.}
\item{sample}{The sample or case ID.}
}
}
\source{
cBioPortal \url{}).
}
\usage{
HER2_data_se2
}
\description{
A gene expression matrix for 20 samples (10 HER2 positive and 10 HER2 negative).
}
\keyword{datasets}
