% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HER2_GEO_microarr}
\alias{HER2_GEO_microarr}
\title{HER2 microarray data set obtained from GEO}
\format{
A matrix containing 21,656 HUGO gene symbols (row names) and 60
breast cancer tumour samples IDs (columns) which are given as sample
accession numbers.
}
\source{
GSE31448 series link on GEO:
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE31448}).
}
\usage{
HER2_GEO_microarr
}
\description{
A matrix containing microarray data for 60 human breast tumour samples (30
samples were ERBB2 (HER2) positive and 30 samples HER2 negative) extracted
from GSE31448 dataset (Sabatier et al. 2011) using the GEO query library.
Multi-gene probes were excluded and only cancer samples from human breast
cancer tumours were selected for constructing this matrix.
}
\references{
Sabatier, R., Finetti, P., Adelaide, J., Guille, A., Borg, J.P.,
Chaffanet, M., Lane, L., Birnbaum, D. and Bertucci, F., 2011. Down-regulation
of ECRG4, a candidate tumor suppressor gene, in human breast cancer.
PloS one, 6(11), p.e27656. doi: https://doi.org/10.1371/journal.pone.0027656
}
\keyword{datasets}
