% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HER2_sig_df}
\alias{HER2_sig_df}
\title{HER2 gene signature}
\format{
A data frame with 387 rows and 2 variables:
\describe{
\item{gene}{The name of highly differentially expressed gene when HER2
pathway is active.}
\item{expression}{The change in expression of the gene when the pathway
is active i.e. -1 for down-regulated genes and 1 for up-regulated genes.}
}
}
\source{
MSigDB \url{https://www.gsea-msigdb.org/gsea/msigdb/geneset_page.jsp?geneSetName=ERBB2_UP.V1_UP}

MSigDB \url{https://www.gsea-msigdb.org/gsea/msigdb/cards/ERBB2_UP.V1_DN.html}
}
\usage{
HER2_sig_df
}
\description{
A list of up-regulated and down-regulated genes constituting the gene
signature of ERBB2 pathway activation. The list of differentially expressed
genes were acquired from MCF-7 cells (breast cancer cell line)
which were ESR1 (estrogen receptor) positive and engineered to express
ligand-activatable ERBB2.
}
\references{
Creighton CJ, Hilger AM, Murthy S, Rae JM, Chinnaiyan AM,
El-Ashry D. Activation of mitogen-activated protein kinase in estrogen
receptor alpha-positive breast cancer cells in vitro induces an in vivo
molecular phenotype of estrogen receptor alpha-negative human breast tumors.
Cancer Res. 2006 Apr 1;66(7):3903-11. doi: 10.1158/0008-5472.CAN-05-4363.
PMID: 16585219.
}
\keyword{datasets}
