% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HER2_Smid_sig}
\alias{HER2_Smid_sig}
\title{HER2 gene signature}
\format{
A data frame with 156 rows (gene names) and 2 variables:
\describe{
\item{gene}{The HUGO gene symbols of a highly differentially expressed gene
when HER2 pathway is active.}
\item{expression}{The change in expression of the gene when the pathway
is active i.e. -1 for down-regulated genes and 1 for up-regulated genes.}
}
}
\source{
MSigDB (For up-regulated component of the:
\url{https://www.gsea-msigdb.org/gsea/msigdb/cards/SMID_BREAST_CANCER_ERBB2_UP.html})

MSigDB (For down-regulated component of the signature:
\url{https://www.gsea-msigdb.org/gsea/msigdb/cards/SMID_BREAST_CANCER_ERBB2_DN.html}
)
}
\usage{
HER2_Smid_sig
}
\description{
A list of up-regulated and down-regulated genes of the ERBB2 (HER2) sub type
(which is characterized by high expression of ERBB2) of 344 primary breast
tumors from lymph node-negative patients (Smid et al. 2008).
}
\references{
Smid, M., Wang, Y., Zhang, Y., Sieuwerts, A.M., Yu, J., Klijn, J.G., Foekens,
J.A. and Martens, J.W., 2008. Subtypes of breast cancer show preferential
site of relapse. Cancer research, 68(9), pp.3108-3114. doi:
\url{https://doi.org/10.1158/0008-5472.CAN-07-5644}
}
\keyword{datasets}
