% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HER2_TCGA_RNA}
\alias{HER2_TCGA_RNA}
\title{HER2 RNA-seq gene expression data set from TCGA}
\format{
A matrix with 20,124 HUGO gene names (rows) and 60 TCGA barcodes which act as
IDs for each tumour sample (columns).
}
\source{
The Cancer Genome Atlas (TCGA) \url{https://portal.gdc.cancer.gov}).
}
\usage{
HER2_TCGA_RNA
}
\description{
A gene expression matrix containing RNA-seq raw read counts for 60 primary
breast tumours (30 tumours of which were human epidermal growth factor
receptor (HER2) positive and 30 of which were HER2 negative). This data set
is a subset of a much larger data set collected from TCGA containing gene
expression data for 1,101 primary breast cancer tumours.
}
\keyword{datasets}
