% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation_script_new.R
\name{log_cpm}
\alias{log_cpm}
\title{Transformation of the raw count data by using  log CPM (counts per million)}
\usage{
log_cpm(data_es)
}
\arguments{
\item{data_es}{Gene expression matrix with gene IDs/symbols as row names
and Sample IDs as column names}
}
\value{
Transformed gene expression data matrix
}
\description{
Transformation of the raw count data by using  log CPM (counts per million) by calling
cpm method in edgeR library
}
\examples{
log_cpm(data_es)

}
\author{
Rishabh Kaushik and Taniya Pal \email{rishabh.kaushik.126@cranfield.ac.uk, taniya.pal.094@cranfiled.ac.uk}
}
