% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_pre-processing.R
\name{log_cpm_transform}
\alias{log_cpm_transform}
\title{Log CPM transformation of RNA-seq raw count data by using log CPM}
\usage{
log_cpm_transform(dataset, boxplot = TRUE)
}
\arguments{
\item{dataset}{An unnormalised gene expression matrix containing raw RNA-seq
(integer) counts with gene IDs/symbols as row names and sample IDs as column
names.}

\item{boxplot}{Optional argument that is a boolean (TRUE or FALSE) indicating
whether boxplots displaying before and after log CPM transformation should be
displayed. (Default=TRUE)}
}
\value{
A logCPM transformed gene expression data matrix with gene symbols /
IDs as row names and samples names / IDs as column names.
}
\description{
Performs a logCPM transformation of RNA-seq raw count data
using the counts per million (CPM) method from the edgeR library. In addition
to the log CPM transformation, the function can also plot a boxplot as sanity
check for logCPM transformation of the gene expression matrix. The first
series of boxplots display the distribution of the raw counts for each sample
in thousands, while the second series of boxplots show the distribution of
the logCPM normalised gene expression matrix.
}
\examples{
\dontrun{log_cpm_transform(data.matrix)}
}
\author{
Rishabh Kaushik and Taniya Pal
\email{rishabh.kaushik.126@cranfield.ac.uk, taniya.pal.094@cranfiled.ac.uk}
}
