% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cpm_transformation.R
\name{log_cpm_transformation}
\alias{log_cpm_transformation}
\title{Transformation of the raw count data by using  log CPM (counts per million) along
with sanity check for transformation}
\usage{
log_cpm_transformation(data_es)
}
\arguments{
\item{data_es}{Gene expression matrix with gene IDs/symbols as row names
and Sample IDs as column names (return value of read_expression matrix function)}
}
\value{
Transformed gene expression data matrix
}
\description{
Transformation of the raw count data by using  log CPM (counts per million)
by calling cpm method in edgeR library. Along with performing transformation
of raw data,the function plots the distribution of raw data before transformation
in the form of a boxplot. Then, it normalizes the raw counts of gene
expression with the log cpm transformation method and returns a boxplot
of the gene expression matrix after transformation for sanity check of the
transformation. The user can check the distribution of the gene expression values
after log cpm transformation with the help of the box plot.
}
\examples{
\dontrun{log_cpm_transformation(formatted_matrix)}
}
\author{
Rishabh Kaushik and Taniya Pal \email{rishabh.kaushik.126@cranfield.ac.uk, taniya.pal.094@cranfiled.ac.uk}
}
