% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{read_expression_data}
\alias{read_expression_data}
\title{Reading gene expression data from file}
\usage{
read_expression_data(file_name)
}
\arguments{
\item{file_name}{Full path of the gene expression data file name}
}
\value{
Structured matrix containing gene symbols/IDs as rownames and
sample IDs as column names.

and the Sample IDs as colnames.
}
\description{
Reads gene expression data file from user,
removes Nas, if the first column consist of gene names,
gives the genenames to the rownames of the matrix and deletes
the first column. Before converting the first column to rownames
it checks for any duplicated gene present in the first column.
Following that the matrix is converted to a numeric data matrix.
}
\examples{
read_expression_data("/Users/taniyapal/Documents/Group Project/TCGA_unannotated.txt")
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
