% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_signature}
\alias{read_signature}
\title{Reads up-regulated and down-regulated gene signatures from gene set files}
\usage{
read_signature(up_sig_file, down_sig_file)
}
\arguments{
\item{up_sig_file}{Up-regulated gene-set format file}

\item{down_sig_file}{Down-regulated gene-set format file}
}
\value{
A data frame containing both up regulated and down
regulated signature files and signifying their up or down
expression with +1 and -1 respectively.
}
\description{
Reads up and down regulated signature files provided either in
gene set file format (.grp) or gene matrix transposed format (.gmt) creating
a data frame which the first column named "gene" containing the gene symbols
and the second column called "expression" containing the corresponding
expression value for a gene in the gene signature, where 1 signifies
up-regulation and -1 represents down-regulation of the gene in the gene
signature.
}
\examples{
\dontrun{read_sign_data("ESR1_UP.v1._UP.csv","ESR1_DN.v1_DN.csv" )}
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
