% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{read_signature_data}
\alias{read_signature_data}
\title{Reads up-regulated and down-regulated gene signatures from files,
formats it according to the needs of the package analysis,
and combines the up and down regulated signature files
into a single data frame}
\usage{
read_signature_data(up_sig_file, down_sig_file)
}
\arguments{
\item{up_sig}{Up-regulated gene-set}

\item{down_sig}{Down-regulated gene-set}
}
\value{
A dataframe containing both up regulated and down
regulated signature files and signifying their up or down
expression with +1 and -1 respetively.
}
\description{
Reads up and down regulated signature files from user
and structures it according to package requirements. It returns a
dataframe. The first column of the dataframe  lists together
both up and down regulated signatures. The second column
signifies whether they are up(+1) or down(-1) regulated.
}
\examples{
read_signature_data("ESR1_UP.v1_UP.csv","ESR1_DN.v1_DN.csv" )
}
\author{
Tania Pal \email{taniya.pal.094@cranfield.ac.uk}
}
