% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{readmatrix}
\alias{readmatrix}
\title{Reading gene expression data matrix}
\usage{
readmatrix(expdata, rowData, colData)
}
\arguments{
\item{matrix}{Gene expression data matrix}
}
\value{
Structured matrix for the convenience of  the package
}
\description{
Reads gene expression matrix from user and structures it according to page requirements
}
\examples{
readmatrix(matrix)
}
\author{
Tania Pal \email{taniya.pal.094@cranfield.ac.uk}
}
