% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{readsig}
\alias{readsig}
\title{Reading up-regulated and down-regulated gene signatures}
\usage{
readsig(up_sig, down_sig)
}
\arguments{
\item{up_sig}{Up-regulated gene-set}

\item{down_sig}{Down-regulated gene-set}
}
\value{
Up and down regulated signature files in the form a list for the convenience of  the package
}
\description{
Reads up and down regulated signature files from user and structures it according to package requirements
}
\examples{
readsig(up_sig, down_sig)
}
\author{
Tania Pal \email{taniya.pal.094@cranfield.ac.uk}
}
