% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation_script_new.R
\name{transform}
\alias{transform}
\title{Transforms gene expression dataset according to CPM read count normalization}
\usage{
transform(data_es)
}
\arguments{
\item{data_es}{Gene expression data matrix}
}
\value{
Normalized Gene expression data matrix
}
\description{
Transforms gene expression dataset according to CPM read count normalization,
calls cmp method in edgeR library
}
\examples{
transform(data_es)

}
\author{
Rishabh Kaushik \email{rishabh.kaushik.126@cranfield.ac.uk}
}
