% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{transform_matrix}
\alias{transform_matrix}
\title{Transformation of gene expression data with log cpm
transformation method}
\usage{
transform_matrix(input)
}
\arguments{
\item{input}{expression matrix after being
pre processed by read_input_file function}
}
\value{
Two boxplots, one representing
distribution of gene expression values before transformation,
another one after transformation.
}
\description{
Plots the raw data before transformation in the form
of a boxplot. Then, it normalizes the raw counts of gene
expression with the log cpm transformation
method and returns a boxplot of the gene expression
matrix after transformation for sanity check of the
transformation. The user can check the distribution of
the gene expression values after log cpm
transformation with the help  of a box plot.
}
\examples{
transform_matrix(input)
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
