% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{visualise_GSVA}
\alias{visualise_GSVA}
\title{GSVA score density plot}
\usage{
visualise_GSVA(sig_df, data_se)
}
\arguments{
\item{sig_df}{Gene expression signature for a specific pathway given as data frame
containing a list of genes that are the most differentially expressed when
a specific pathway is active. Up-regulated genes are given an
expression of 1, while down-regulated genes are given an expression of -1.}

\item{data_se}{Expression data set matrix given as a matrix or data frame
comprising the expression levels of genes in a set of samples. Gene
expression matrices can be from RNASeq
or microarray transcriptomic data sets.}
}
\value{
density plot displaying distribution of GSVA scores obtainied for
the samples using up-regulated and down-regulated gene-sets from the gene
signature
}
\description{
Plots GSVA scores distribution for up-regulated and
down-regulated gene-sets of a provided gene expression signature
}
\examples{
visualise_GSVA(ER_sig_df, ER_data_se)
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
