% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_accuracy.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Accuracy calculation of classification method}
\usage{
calculate_accuracy(
  true_labels_source,
  predicted_labels_source,
  pathway,
  display_statistics = FALSE,
  display_roc_curve = FALSE
)
}
\arguments{
\item{true_labels_source}{data frame ,matrix or name of file which contains sample Ids and corresponding actual labels.}

\item{predicted_labels_source}{predicted labels data frame or matrix generated by classification method}

\item{pathway}{name of pathway, currently HER2 and ER are defined.}

\item{display_statistics}{optional flag to display statistical information about confusion matrix}

\item{display_roc_curve}{optional flag to plot ROC Curve for confusion matrix}
}
\value{
confusion_matrix
}
\description{
Accuracy calculation is performed using classified labels from the classification method and actual labels.
Confusion matrix is created to display the accuracy to the user. Additional statistical information and ROC curve are
optional features. It should be called after classification method to evaluate accuracy of classification
}
\examples{
\dontrun{calculate_accuracy(true_labels_df,predicted_labels_df, "ER",
display_statistics= TRUE, display_roc_curve=TRUE)}
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
