% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_accuracy.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Classifies samples according to pathway activity using GSVA
Creates confusion matrix and calculate accuracy of classification method}
\usage{
calculate_accuracy(true_labels_df, predicted_labels_df)
}
\arguments{
\item{true_labels_df}{actual labels dataframe or matrix}

\item{predicted_labels_df}{predicted labels dataframe or matrix generated in classification method}
}
\value{
confusion_matrix_HER2
}
\description{
Classifies samples according to pathway activity using GSVA
Creates confusion matrix and calculate accuracy of classification method
}
\examples{
calculate_accuracy(HER2_labelDF,Her2_pathway)
}
\author{
Ozlem Karadeniz \email{ozlem.karadeniz.283@cranfield.ac.uk}
}
