% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_pca.R
\name{classes_pca}
\alias{classes_pca}
\title{pca plot for normalized gene expression dataset}
\usage{
classes_pca(normalized_data_filtered, predicted_labels_df, pathway_name = "ER")
}
\arguments{
\item{normalized_data_filtered}{Normalized Gene expression data matrix}

\item{predicted_labels_df}{dataframe which labels for each sample predicted by classification algorithm}

\item{pathway_name}{pathway name, default is "ER"}
}
\value{

}
\description{
plots pca for normalized gene expression dataset with the predicted labels
}
\examples{
\dontrun{classes_pca(normalized_data_filtered, predicted_labels_df, 'ER')}
}
\author{
Yi-Hsuan Lee \email{yi-hsuan.lee@cranfield.ac.uk}
}
