% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classifies samples according to pathway activity using GSVA}
\usage{
classify(sigDf, dataSe, thresh = 0.25)
}
\arguments{
\item{sigDf}{Signature dataframe}

\item{dataSe}{Expression dataset matrix}

\item{thresh}{Threshold between 0-1 for pathway classification (default: 0.25)}
}
\value{
classesDf
}
\description{
Classifies samples according to pathway activity using GSVA
}
\examples{
# Default threshold
classify(HER2_sig_df, HER_data_se)
# Threshold = 0.05 (top 5\% of up-regulated and bottom 5\% of down-regulated genes)
classify(HER2_sig_df, HER_data_se, thresh=0.05)
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
