% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify}
\alias{classify}
\title{Generic classify function for classifying samples based on GSVA scores}
\usage{
classify(
  gsva_scores,
  up_thresh.low,
  up_thresh.high,
  dn_thresh.low,
  dn_thresh.high
)
}
\arguments{
\item{gsva_scores}{List containing a data frame (Up) of samples and their
corresponding GSVA scores when running GSVA on the up-regulated gene-set, and
a dataframe (Down) with GSVA scores for the same samples when running GSVA
on the down-regulated gene-set.}

\item{up_thresh}{Numerical vector containing the low and high threshold for
classification when checking consistency with only the up-regulated gene-set
of the gene signature}

\item{dn_thresh}{Numerical vector containing the low and high threshold for
classification when checking consistency with only the down-regulated
gene-set of the gene signature}
}
\value{
data frame containing the pathway activity class labels for each
sample i.e. "Active", "Inactive" or "Uncertain".
}
\description{
Generic classify function for classifying samples based on GSVA scores
}
\examples{
classes_df <- classify(gsva_scores, up_thresh, dn_thresh)
}
