% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_gsva_abs}
\alias{classify_gsva_abs}
\title{Classification Using Absolute GSVA Score Thresholds}
\usage{
classify_gsva_abs(
  expr_mat,
  sig_df,
  up_thresh.low,
  up_thresh.high,
  dn_thresh.low,
  dn_thresh.high
)
}
\arguments{
\item{expr_mat}{Normalised expression data set matrix comprising the
expression levels of genes (rows) for each sample (columns) in a data set.
Row names are gene symbols and column names are sample IDs / names. Gene
expression matrices can contain normalised (logCPM transformed) RNASeq or
microarray transcriptomic data.}

\item{sig_df}{Gene expression signature for a specific pathway given as data
frame with the first column named "gene" containing a list of genes that are
the most differentially expressed when the given pathway is active and the
second column named "expression" containing their corresponding expression:
-1 for down-regulated genes and 1 for up-regulated genes.}

\item{up_thresh.low}{Number denoting the absolute GSVA score threshold for
categorizing a sample as having inconsistent expression with the up-regulated
gene-set from the gene signature.}

\item{up_thresh.high}{Number denoting the absolute GSVA score threshold for
categorizing a sample as having consistent expression with the up-regulated
gene set from the gene signature.}

\item{dn_thresh.low}{Number denoting the absolute GSVA score threshold for
categorizing a sample as having consistent expression with the down-regulated
gene-set from the gene signature.}

\item{dn_thresh.high}{Number denoting the absolute GSVA score threshold for
categorizing a sample as having inconsistent expression with the
down-regulated gene-set from the gene signature.}
}
\value{
A data frame containing a list of samples as the first column and
their classified pathway activity (Active, Inactive or Uncertain) in the
second column.
}
\description{
Classifies samples according to pathway activity by first
ranking samples by their GSVA score and assessing evidence of expression
consistency of each sample with the up-regulated gene-set and down-regulated
gene-set of the gene signature using absolute GSVA score thresholds. GSVA
scores generated by the GSVA algorithm provide a measure of expression
abundance of the up-regulated and down-regulated gene-sets, which
PathAnalyser uses to assess expression consistency with both
parts (up-regulated and down-regulated gene-sets) of the gene signature using
the user-supplied absolute GSVA score thresholds as thresholds for expression
consistency with each part of the signature.
}
\details{
Four thresholds are specified by the user:
\enumerate{
\item "up_thresh.low" - a GSVA score threshold for considering a sample as
having inconsistent expression with up-regulated gene-set.
\item "up_thresh.high" - a GSVA score threshold for considering a sample as
having expression consistent with the up-regulated gene-set up-regulated gene
set
\item "dn_thresh.low" - a GSVA score threshold for considering a sample as
having consistent expression with down-regulated gene-set
\item "dn_thresh.high" - a GSVA score threshold for considering a sample as
having inconsistent expression with the down-regulated gene-set.
Samples that have consistent expression with both the up-regulated and
down-regulated gene-sets of the signature are classified as "Active", those
with inconsistent expression with both parts of the signature are classified
as "Inactive" and the rest of the samples are classified as "Uncertain".
}
}
\examples{
# Default thresholds for up-regulated and down-regulated gene-sets
\dontrun{classes_df <- classify_gsva_abs(ER_dataset, ER_sig, up_thresh.low=-0.25,
     up_thresh.high=0.25, dn_thresh.low=-0.25, dn_thresh.high=0.35)}
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
