% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_GSVA_percent}
\alias{classify_GSVA_percent}
\title{Sample classification according to pathway activity using a percentile
threshold for assessing expression consistency with a gene signature.}
\usage{
classify_GSVA_percent(sig_df, data_se, percent_thresh = 25)
}
\arguments{
\item{sig_df}{Gene expression signature for a specific pathway given as data
frame containing a list of genes that are the most differentially expressed
when a specific pathway is active. Up-regulated genes are given an
expression of 1, while down-regulated genes are given an expression of -1.}

\item{data_se}{Normalized gene expression matrix or data frame containing
expression values for samples in a data set (RNA Seq read counts or
microarray data).}

\item{percent_thresh}{Percentile threshold (0-100) of samples for checking
consistency of gene expression of a sample with first the up-regulated and
then down-regulated gene-set of the gene signature (default= 25\% (quartile)).}
}
\value{
data frame containing predicted pathway activity classes for each
sample in the user provided data set (Active, Inactive or Uncertain).
}
\description{
Classifies samples by ranking samples by abundance of expression
calculated via the GSVA algorithm. Samples are assessed for expression
consistency with both the up-regulated and down-regulated gene-sets using
percentile thresholds during the pathway activity sample classification.
}
\examples{
# default using quartile threshold (25th percentile)
classes_df <- classify_GSVA_percent(ER_sig_df, ER_data_se1)
# custom percentile threshold e.g. 30th percentile
classes_df <- classify_GSVA_percent(ER_sig_df, ER_data_se1,
       thresh_percent=30)
}
\author{
Anisha Thind \email{a.thind@cranfield.ac.uk}
}
