% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ER_data_se1}
\alias{ER_data_se1}
\title{ER gene expression data set 1}
\format{
A data frame with 20,124 rows and 20 variables:
\describe{
\item{gene}{The name of the gene for which expression data is provided.}
\item{sample}{The sample or case ID.}
}
}
\source{
\url{https://portal.gdc.cancer.gov}).
}
\usage{
ER_data_se1
}
\description{
Raw counts data obtained from TCGA. The data set was filtered for Breast cancer-primary tumors.
Count data subset: A gene expression matrix containing raw read counts from RNA-seq for 20
samples (10 ER positive and 10 ER negative).
}
\keyword{datasets}
