% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ER_sig_df}
\alias{ER_sig_df}
\title{ER gene signature}
\format{
A data frame with 160 rows and 2 variables:
\describe{
\item{gene}{The name of highly differentially expressed gene when ER
pathway is active.}
\item{expression}{The change in expression of the gene when the pathway
is active i.e. -1 for down-regulated genes and 1 for up-regulated genes.}
}
}
\source{
SET index defined by Symmans et al. 2010 (see
\url{https://pubmed.ncbi.nlm.nih.gov/20697068/}).
(PMID: 20697068
PMCID: PMC2953969 DOI: 10.1200/JCO.2010.28.4273)
}
\usage{
ER_sig_df
}
\description{
A list of up-regulated and down-regulated genes constituting the gene
signature of ER pathway activation.
}
\references{
Symmans, W.F., Hatzis, C., Sotiriou, C., Andre, F., Peintinger,
F., Regitnig, P., Daxenbichler, G., Desmedt, C., Domont, J., Marth, C. and
Delaloge, S., 2010. Genomic index of sensitivity to endocrine therapy for
breast cancer. Journal of clinical oncology, 28(27), p.4111.
}
\keyword{datasets}
