% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation_script_new.R
\name{log_cpm}
\alias{log_cpm}
\title{Normalisation of the raw count data by using  log CPM (counts per million)}
\usage{
log_cpm(data_es)
}
\arguments{
\item{data_es}{Gene expression data matrix from TCGA}
}
\value{
Normalized Gene expression data matrix
}
\description{
Normalisation of the raw count data by using  log CPM (counts per million) by calling
cmp method in edgeR library
}
\examples{
log_cpm(data_es)

}
\author{
Rishabh Kaushik \email{rishabh.kaushik.126@cranfield.ac.uk}
}
