% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cpm_transformation.R
\name{log_cpm_transform}
\alias{log_cpm_transform}
\title{Log CPM transformation of RNA-seq raw count data by using log CPM}
\usage{
log_cpm_transform(data_es)
}
\arguments{
\item{data_es}{An unnormalised gene expression matrix containing raw RNA-seq
counts with gene IDs/symbols as row names and sample IDs as column names}
}
\value{
logCPM transformed gene expression data matrix
}
\description{
Performs a logCPM transformation of RNA-seq raw count data
using the counts per million (CPM) method from edgeR library. In addition to
the log CPM transformation, the function plots two boxplots as sanity check
for logCPM transformation of a gene expression matrix. The first boxplot
displays the distribution of the raw counts for each sample in thousands,
while the second boxplot shows the distribution of the logCPM normalised gene
expression matrix.
}
\examples{
\dontrun{log_cpm_transformation(formatted_matrix)}
}
\author{
Rishabh Kaushik and Taniya Pal \email{rishabh.kaushik.126@cranfield.ac.uk, taniya.pal.094@cranfiled.ac.uk}
}
