% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50}
\alias{pam50}
\title{Gene signature for ER and HER2 extracted from PAM50}
\format{
List of two of two data frames: one for ER signature, the other for
HER2 signature:
\describe{
\item{ER}{data frame containing gene names and their relative expression
value (-1 for down-regulated genes and 1 for up-regulated genes) in the
ER signature}
\item{HER2}{data frame containing gene names their relative expression
value (-1 for down-regulated genes and 1 for up-regulated genes) in the
HER2 signature}
}
}
\usage{
pam50
}
\description{
A list of two data frames containing the gene names and their corresponding
expression values constituting ER and HER gene signatures obtained from
PAM50.
}
\keyword{datasets}
