% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_expression_data}
\alias{read_expression_data}
\title{Reading gene expression data from file}
\usage{
read_expression_data(file)
}
\arguments{
\item{file}{Path for gene expression matrix file, which is either tab or
comma value separated. The first column must contain gene names and the
subsequent column names should be the sample name or ID.}
}
\value{
Structured matrix containing gene symbols/IDs as row names and
sample IDs as column names and the Sample IDs as column names.
}
\description{
Reads gene expression matrix data file from user,
removes NAs, if the first column consist of gene names,
gives the gene names to the row names of the matrix and deletes
the first column. Before converting the first column to row names
it checks for any duplicated gene present in the first column.
Following that the matrix is converted to a numeric data matrix.
}
\examples{
\dontrun{read_expr_data("/Users/taniyapal/Documents/Group Project/TCGA_unannotated.txt")}
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
