% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_expression_data}
\alias{read_expression_data}
\title{Reading gene expression data from file}
\usage{
read_expression_data(file)
}
\arguments{
\item{file}{Path for gene expression matrix file, which is either tab /
comma / white-space value separated. The first column must contain gene
symbols and the subsequent column names should be the sample name or ID.}
}
\value{
A numerical matrix containing gene symbols/IDs as row names and
sample IDs as column names.
}
\description{
Reads gene expression matrix data file which is either
tab/comma/white-space value separated text file. After reading in the input
file, any gene rows that contain NAs are removed and the data set is screened
for duplicates. Duplicate genes are reduced to single gene row entries, with
each value corresponding to the mean expression value for each sample for the
duplicated gene entry. If duplicate samples are detected then the only the
first sample of the duplicates are retained. Finally, the data frame is
converted to a numerical matrix, where row names represent gene symbols and
columns represent sample names or IDs.
}
\examples{
\dontrun{read_expr_data("/Users/taniyapal/Documents/Group Project/TCGA_unannotated.txt")}
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
