% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_matrix_evaluation.R
\name{read_input_file}
\alias{read_input_file}
\title{Reading both expression data and signature files,also checking overlap
between matrix and signature gene symbols}
\usage{
read_input_file(
  input_filename,
  up_sig_filename,
  dn_sig_filename,
  check_matrix_sig_overlap = F
)
}
\arguments{
\item{input_filename}{The expression matrix file name}

\item{up_sig_filename}{The up regulated signature data file name}

\item{dn_sig_filename}{The down regulated signature data file name}

\item{check_matrix_sig_overlap}{Boolean, if true returns a Venn diagram plotting
overlap between matrix and gene signature file symbols/IDs.}
}
\value{
A list with 3 objcets:
\enumerate{
\item formatted expression matrix with gene IDs/symbols as row names and
samples IDs as column names
\item a data frame, whose first column has combined up and down regulated signature
IDs/symbols, second column has representation of up regulated expression as +1
and down regulated expression as -1
\item a Venn diagram representing the overlap of gene symbols/IDs between the
two files
}
}
\description{
Reads both expression data and up and down regulated signature
data, if check_matrix_sig_overlap parameter in the function is true, returns
a Venn diagram representating which gene symbols
overlap between the signature file
and the expression matrix file.
The user gets an estimates of how many gene symbols overlap
between the two files.
}
\examples{
read_input_file("TCGA_unannotated.txt","ESR1_UP.v1_UP.csv",
"ESR1_DN.v1_DN.csv", check_mtarix_sig_overlap=T)
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
