% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_and_matrix_evaluation_functions.R
\name{read_input_file}
\alias{read_input_file}
\title{Reading gene expression data from file}
\usage{
read_input_file(file_name)
}
\arguments{
\item{matrix}{Full path of the gene expression data file name}
}
\value{
Structured matrix containing gene symbols/IDs as rownames
Also returns a heatmap with identifiable duplicated samples for
the convinience of the user.
and the Sample IDs as colnames.
}
\description{
Reads gene expression data file from user,
removes Nas, if the first column consist of gene names,
gives the genenames to the rownames of the matrix and deletes
the first column. Before converting the first column to rownames
it checks for any duplicated gene present in the first column.
Following that the matrix is converted to a numeric data matrix.
It also returns a heatmap which helps the user
to identify duplicate samples (columns) from the
heatmap, which in turn will help the user to remove duplicated
samples according to their choice.
}
\examples{
read_expression_data("/Users/taniyapal/Documents/Group Project/Matrix used for GSVA Analysis by Taniya.csv"))
}
\author{
Tania Pal \email{taniya.pal.094@cranfield.ac.uk}
}
