% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_matrix_evaluation.R
\name{read_signature_data}
\alias{read_signature_data}
\title{Reads up-regulated and down-regulated gene signatures from files}
\usage{
read_signature_data(up_sig_file, down_sig_file)
}
\arguments{
\item{up_sig}{Up-regulated gene-set}

\item{down_sig}{Down-regulated gene-set}
}
\value{
A dataframe containing both up regulated and down
regulated signature files and signifying their up or down
expression with +1 and -1 respetively.
}
\description{
Reads up and down regulated signature files from user
and structures it according to package requirements. It returns a
dataframe. The first column of the dataframe  lists together
both up and down regulated signatures. The second column
signifies whether they are up(+1) or down(-1) regulated.
}
\examples{
read_signature_data("ESR1_UP.v1._UP.csv","ESR1_DN.v1_DN.csv" )
}
\author{
Taniya Pal \email{taniya.pal.094@cranfield.ac.uk}
}
